# feat: faster isNonEvmAddress check